+++ src/main.lua
@@ -45,6 +45,7 @@ end
 local function on_reload()
 	-- what to do when we are ready, but also again on every reload.
 	-- only do things that are safe to run over and over.
+	if config.enabled == false then return end
 	import_as_fallback(rom.game)
 	import 'reload.lua'
 end